refactor: change name for ErrorItem
